+++ src/domain/rules.ts
@@ -1,1 +1,5 @@
 export { validateFinalize } from "./rules/validateFinalize";
+export {
+  MAX_CHECKLIST_ITEMS,
+  canAddChecklistItem,
+} from "./rules/checklistItemLimit";